+++ libs/lwjgl/llvm_license.txt
@@ -16,7 +16,7 @@ The LLVM Project is under the Apache License v2.0 with LLVM Exceptions:
       "Licensor" shall mean the copyright owner or entity authorized by
       the copyright owner that is granting the License.
 
-      "Legal Entity" shall mean the union of the acting entity and all
+      "Legal Logic.Entity" shall mean the union of the acting entity and all
       other entities that control, are controlled by, or are under common
       control with that entity. For the purposes of this definition,
       "control" means (i) the power, direct or indirect, to cause the
@@ -24,7 +24,7 @@ The LLVM Project is under the Apache License v2.0 with LLVM Exceptions:
       otherwise, or (ii) ownership of fifty percent (50%) or more of the
       outstanding shares, or (iii) beneficial ownership of such entity.
 
-      "You" (or "Your") shall mean an individual or Legal Entity
+      "You" (or "Your") shall mean an individual or Legal Logic.Entity
       exercising permissions granted by this License.
 
       "Source" form shall mean the preferred form for making modifications,
@@ -53,7 +53,7 @@ The LLVM Project is under the Apache License v2.0 with LLVM Exceptions:
       the original version of the Work and any modifications or additions
       to that Work or Derivative Works thereof, that is intentionally
       submitted to Licensor for inclusion in the Work by the copyright owner
-      or by an individual or Legal Entity authorized to submit on behalf of
+      or by an individual or Legal Logic.Entity authorized to submit on behalf of
       the copyright owner. For the purposes of this definition, "submitted"
       means any form of electronic, verbal, or written communication sent
       to the Licensor or its representatives, including but not limited to
@@ -63,7 +63,7 @@ The LLVM Project is under the Apache License v2.0 with LLVM Exceptions:
       excluding communication that is conspicuously marked or otherwise
       designated in writing by the copyright owner as "Not a Contribution."
 
-      "Contributor" shall mean Licensor and any individual or Legal Entity
+      "Contributor" shall mean Licensor and any individual or Legal Logic.Entity
       on behalf of whom a Contribution has been received by Licensor and
       subsequently incorporated within the Work.
