+++ src/Logic/DirectionalLight.java
@@ -1,38 +1,23 @@
+package Logic;
 
-public class SpotLight extends Light
+public class DirectionalLight extends Light
 {
-	
-	public SpotLight()
+	public DirectionalLight()
 	{
 		super();
 	}
 	
-	public SpotLight(float directionX, float directionY, float directionZ,
-					 float positionX, float positionY, float positionZ,
-					 float colorR, float colorG, float colorB,
-					 float angleRadians)
+	public DirectionalLight(float directionX, float directionY, float directionZ,  float colorR, float colorG, float colorB)
 	{
 		super(colorR, colorG, colorB);
 		this.directionX = directionX;
 		this.directionY = directionY;
 		this.directionZ = directionZ;
-		
-		this.positionX = positionX;
-		this.positionY = positionY;
-		this.positionZ = positionZ;
-		
-		this.angleCos = (float)Math.cos(angleRadians);
 	}
 	
 	public float directionX = 0;
 	public float directionY = 0;
 	public float directionZ = 0;
-	
-	public float positionX = 0;
-	public float positionY = 0;
-	public float positionZ = 0;
-	
-	public float angleCos = 0;
 	
 	void normalizeData()
 	{
@@ -50,6 +35,4 @@ public class SpotLight extends Light
 			directionZ *= inverseSrt;
 		}
 	}
-	
-	
 }
