+++ src/Logic/IntVector.java
@@ -1,3 +1,5 @@
+package Logic;
+
 public class IntVector
 {
