+++ src/Logic/LightManager.java
@@ -1,3 +1,5 @@
+package Logic;
+
 import org.lwjgl.opengl.GL43;
