+++ src/Logic/Material.java
@@ -1,4 +1,4 @@
-import org.w3c.dom.Text;
+package Logic;
 
 public class Material
 {
