+++ src/Logic/SkyBoxRenderer.java
@@ -1,3 +1,5 @@
+package Logic;
+
 import org.joml.Matrix4f;
 import org.lwjgl.opengl.GL30;
 import org.lwjgl.opengl.GL44;
