+++ src/Logic/SpotLight.java
@@ -1,21 +1,39 @@
-public class DirectionalLight extends Light
+package Logic;
+
+public class SpotLight extends Light
 {
-	public DirectionalLight()
+	
+	public SpotLight()
 	{
 		super();
 	}
 	
-	public DirectionalLight(float directionX, float directionY, float directionZ,  float colorR, float colorG, float colorB)
+	public SpotLight(float directionX, float directionY, float directionZ,
+					 float positionX, float positionY, float positionZ,
+					 float colorR, float colorG, float colorB,
+					 float angleRadians)
 	{
 		super(colorR, colorG, colorB);
 		this.directionX = directionX;
 		this.directionY = directionY;
 		this.directionZ = directionZ;
+		
+		this.positionX = positionX;
+		this.positionY = positionY;
+		this.positionZ = positionZ;
+		
+		this.angleCos = (float)Math.cos(angleRadians);
 	}
 	
 	public float directionX = 0;
 	public float directionY = 0;
 	public float directionZ = 0;
+	
+	public float positionX = 0;
+	public float positionY = 0;
+	public float positionZ = 0;
+	
+	public float angleCos = 0;
 	
 	void normalizeData()
 	{
@@ -33,4 +51,6 @@ public class DirectionalLight extends Light
 			directionZ *= inverseSrt;
 		}
 	}
+	
+	
 }
