+++ src/Logic/StorageBLockBindings.java
@@ -1,3 +1,4 @@
+package Logic;
 
 //used as an enum. I needed an int value associated with it. Used to send data to the GPU.
 public class StorageBLockBindings
