+++ src/Logic/Texture.java
@@ -1,0 +1,14 @@
+package Logic;
+
+public class Texture
+{
+	
+	public int id;
+	
+	public Texture load(String name)
+	{
+		id = TextureLoader.load(name);
+		return  this;
+	}
+
+}
