+++ src/Logic/TextureLoader.java
@@ -1,3 +1,5 @@
+package Logic;
+
 import org.joml.Matrix4f;
 import org.joml.Vector2i;
 import org.joml.Vector3f;
